django-backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.1.6
django-split-settings==1.0.0
djangorestframework==3.11.0
jupyter==1.0.0
psycopg2==2.8.5
notebook==6.0.1
pytz==2019.3
sqlparse==0.3.1
